+++ lib/rdf/serializations/turtle_encoder.ex
@@ -47,7 +47,8 @@ defmodule RDF.Turtle.Encoder do
     end
   end
 
-  defp base_iri(nil, %RDF.Graph{base_iri: base_iri}), do: base_iri
+  defp base_iri(nil, %RDF.Graph{base_iri: base_iri}) when not is_nil(base_iri), do: base_iri
+  defp base_iri(nil, _), do: RDF.default_base_iri()
   defp base_iri(base_iri, _), do: RDF.iri(base_iri)
 
   defp init_base_iri(nil), do: nil