+++ CHANGELOG.md
@@ -18,8 +18,12 @@ This project adheres to [Semantic Versioning](http://semver.org/) and
 ### Changed
 
 - `RDF.Turtle.Decoder` saves the base IRI in the `RDF.Graph` now
-- `RDF.Turtle.Encoder` now takes the base IRI to be serialized from the graph when 
-  no base IRI is given with the `base` option or its new alias `base_iri`
+- `RDF.Turtle.Encoder` now takes the base IRI to be used during serialization in  
+  the following order of precedence:
+	- from the the `base` option or its new alias `base_iri`
+	- from the `base_iri` field of the given graph
+	- from the `RDF.default_base_iri` returning the one from the application 
+	  configuration
 
 
 [Compare v0.6.1...HEAD](https://github.com/marcelotto/rdf-ex/compare/v0.6.1...HEAD)